code refactor for assign function of memory chip
