+++ vm-circuit/src/chips/memory_chip.rs
@@ -43,8 +43,6 @@ pub struct MemoryChipConfig<F: FieldExt> {
     global_op_config: GlobalOpChipConfig<F>,
     s_add_counters: Selector,
     gc_table: TableColumn,
-    frame_index_table: TableColumn,
-    locals_index_table: TableColumn,
 }
 
 #[derive(Clone, Debug)]
@@ -78,16 +76,8 @@ impl<F: FieldExt> MemoryChip<F> {
     pub fn configure(meta: &mut ConstraintSystem<F>) -> <Self as Chip<F>>::Config {
         let advices = [(); MEM_CHIP_WIDTH].map(|_| meta.advice_column());
         let gc_table = meta.lookup_table_column();
-        let frame_index_table = meta.lookup_table_column();
-        let locals_index_table = meta.lookup_table_column();
         let stack_op_config = StackOpChip::configure(meta, advices, &gc_table);
-        let locals_op_config = LocalsOpChip::configure(
-            meta,
-            advices,
-            &gc_table,
-            &frame_index_table,
-            &locals_index_table,
-        );
+        let locals_op_config = LocalsOpChip::configure(meta, advices, &gc_table);
         let global_op_config = GlobalOpChip::configure(meta, advices, &gc_table);
 
         for column in &advices {
@@ -115,8 +105,6 @@ impl<F: FieldExt> MemoryChip<F> {
             global_op_config,
             s_add_counters,
             gc_table,
-            frame_index_table,
-            locals_index_table,
         }
     }
 
@@ -134,222 +122,16 @@ impl<F: FieldExt> MemoryChip<F> {
         let global_ops_num = self.witness.circuit_config.global_ops_num.unwrap_or(0);
 
         let stack_op_chip = StackOpChip::<F>::construct(self.config.stack_op_config.clone(), ());
-        let mut last_stack_counter: Option<AssignedCell<F, F>> = None;
-
-        if !stack_ops.is_empty() || stack_ops_num > 0 {
-            layouter.assign_region(
-                || "stack operations",
-                |mut region: Region<'_, F>| {
-                    let mut counter = 0;
-                    for (index, op) in stack_ops.iter().enumerate() {
-                        counter = index + 1;
-                        let assigned_counter = if index == 0 {
-                            stack_op_chip
-                                .config
-                                .s_first_stack_op
-                                .enable(&mut region, index)?;
-                            stack_op_chip.assign(&mut region, index, op, counter, false)?
-                        } else {
-                            stack_op_chip.config.s_stack_op.enable(&mut region, index)?;
-                            stack_op_chip.assign(&mut region, index, op, counter, false)?
-                        };
-                        if counter == stack_ops.len() {
-                            last_stack_counter = Some(assigned_counter);
-                        }
-                    }
-
-                    // If the number of stack ops is less than stack_ops_num set by user, fill with
-                    // empty op. This happened when the execution path is not fixed, for example,
-                    // if there is loop in the code.
-                    if stack_ops.len() < stack_ops_num {
-                        for index in stack_ops.len()..stack_ops_num {
-                            let assigned_counter = if index == 0 {
-                                stack_op_chip
-                                    .config
-                                    .s_first_stack_op
-                                    .enable(&mut region, index)?;
-                                stack_op_chip.assign(
-                                    &mut region,
-                                    index,
-                                    &ConvertedRWOperation::empty(),
-                                    counter,
-                                    true,
-                                )?
-                            } else {
-                                stack_op_chip.config.s_stack_op.enable(&mut region, index)?;
-                                stack_op_chip.assign(
-                                    &mut region,
-                                    index,
-                                    &ConvertedRWOperation::empty(),
-                                    counter,
-                                    true,
-                                )?
-                            };
-                            last_stack_counter = Some(assigned_counter);
-                        }
-                    }
-                    Ok(())
-                },
-            )?;
-        }
+        let last_stack_counter =
+            stack_op_chip.assign(layouter, circuit_config, stack_ops, stack_ops_num);
 
         let locals_op_chip = LocalsOpChip::<F>::construct(self.config.locals_op_config.clone(), ());
-        let mut last_locals_counter: Option<AssignedCell<F, F>> = None;
-
-        if !locals_ops.is_empty() || locals_ops_num > 0 {
-            layouter.assign_region(
-                || "locals operations",
-                |mut region: Region<'_, F>| {
-                    let mut prev_op = None;
-                    let mut counter = 0;
-                    for (index, op) in locals_ops.iter().enumerate() {
-                        counter = index + 1;
-                        let assigned_counter = if index == 0 {
-                            locals_op_chip
-                                .config
-                                .s_first_locals_op
-                                .enable(&mut region, index)?;
-                            locals_op_chip.assign(&mut region, index, op, counter, None, false)?
-                        } else {
-                            locals_op_chip
-                                .config
-                                .s_locals_op
-                                .enable(&mut region, index)?;
-                            locals_op_chip.assign(
-                                &mut region,
-                                index,
-                                op,
-                                counter,
-                                prev_op,
-                                false,
-                            )?
-                        };
-                        if counter == locals_ops.len() {
-                            last_locals_counter = Some(assigned_counter);
-                        }
-                        prev_op = Some(op.clone());
-                    }
-
-                    // If the number of locals ops is less than locals_ops_num set by user, fill with
-                    // empty locals op.
-                    if locals_ops.len() < locals_ops_num {
-                        for index in locals_ops.len()..locals_ops_num {
-                            let assigned_counter = if index == 0 {
-                                locals_op_chip
-                                    .config
-                                    .s_first_locals_op
-                                    .enable(&mut region, index)?;
-                                locals_op_chip.assign(
-                                    &mut region,
-                                    index,
-                                    &ConvertedRWOperation::empty(),
-                                    counter,
-                                    None,
-                                    true,
-                                )?
-                            } else {
-                                locals_op_chip
-                                    .config
-                                    .s_locals_op
-                                    .enable(&mut region, index)?;
-                                locals_op_chip.assign(
-                                    &mut region,
-                                    index,
-                                    &ConvertedRWOperation::empty(),
-                                    counter,
-                                    prev_op,
-                                    true,
-                                )?
-                            };
-
-                            last_locals_counter = Some(assigned_counter);
-                            prev_op = Some(ConvertedRWOperation::empty());
-                        }
-                    }
-
-                    Ok(())
-                },
-            )?;
-        }
+        let last_locals_counter =
+            locals_op_chip.assign(layouter, circuit_config, locals_ops, locals_ops_num);
 
         let global_op_chip = GlobalOpChip::<F>::construct(self.config.global_op_config.clone(), ());
-        let mut last_global_counter: Option<AssignedCell<F, F>> = None;
-
-        if !global_ops.is_empty() || global_ops_num > 0 {
-            layouter.assign_region(
-                || "global operations",
-                |mut region: Region<'_, F>| {
-                    let mut prev_op = None;
-                    let mut counter = 0;
-                    for (index, op) in global_ops.iter().enumerate() {
-                        counter = index + 1;
-                        let assigned_counter = if index == 0 {
-                            global_op_chip
-                                .config
-                                .s_first_global_op
-                                .enable(&mut region, index)?;
-                            global_op_chip.assign(&mut region, index, op, counter, None, false)?
-                        } else {
-                            global_op_chip
-                                .config
-                                .s_global_op
-                                .enable(&mut region, index)?;
-                            global_op_chip.assign(
-                                &mut region,
-                                index,
-                                op,
-                                counter,
-                                prev_op,
-                                false,
-                            )?
-                        };
-                        if counter == global_ops.len() {
-                            last_global_counter = Some(assigned_counter);
-                        }
-                        prev_op = Some(op.clone());
-                    }
-
-                    // If the number of global ops is less than global_ops_num set by user, fill with
-                    // empty locals op.
-                    if global_ops.len() < global_ops_num {
-                        for index in global_ops.len()..global_ops_num {
-                            let assigned_counter = if index == 0 {
-                                global_op_chip
-                                    .config
-                                    .s_first_global_op
-                                    .enable(&mut region, index)?;
-                                global_op_chip.assign(
-                                    &mut region,
-                                    index,
-                                    &ConvertedRWOperation::empty(),
-                                    counter,
-                                    None,
-                                    true,
-                                )?
-                            } else {
-                                global_op_chip
-                                    .config
-                                    .s_global_op
-                                    .enable(&mut region, index)?;
-                                global_op_chip.assign(
-                                    &mut region,
-                                    index,
-                                    &ConvertedRWOperation::empty(),
-                                    counter,
-                                    prev_op,
-                                    true,
-                                )?
-                            };
-
-                            last_global_counter = Some(assigned_counter);
-                            prev_op = Some(ConvertedRWOperation::empty());
-                        }
-                    }
-
-                    Ok(())
-                },
-            )?;
-        }
+        let last_global_counter =
+            global_op_chip.assign(layouter, circuit_config, global_ops, global_ops_num);
 
         layouter.assign_region(
             || "add counter",
@@ -476,38 +258,6 @@ impl<F: FieldExt> MemoryChip<F> {
                 }
 
                 Ok(())
-            },
-        )?;
-
-        layouter.assign_table(
-            || "frame_index_table",
-            |mut table_column| {
-                (0..=circuit_config.max_frame_index)
-                    .map(|i| {
-                        table_column.assign_cell(
-                            || format!("frame_index_table[{}]", i),
-                            self.config.frame_index_table,
-                            i,
-                            || CircuitValue::known(F::from_u128(i as u128)),
-                        )
-                    })
-                    .fold(Ok(()), |acc, res| acc.and(res))
-            },
-        )?;
-
-        layouter.assign_table(
-            || "locals_index_table",
-            |mut table_column| {
-                (0..=circuit_config.max_locals_size)
-                    .map(|i| {
-                        table_column.assign_cell(
-                            || format!("locals_index_table[{}]", i),
-                            self.config.locals_index_table,
-                            i,
-                            || CircuitValue::known(F::from_u128(i as u128)),
-                        )
-                    })
-                    .fold(Ok(()), |acc, res| acc.and(res))
             },
         )?;
