+++ vm-circuit/src/chips/memory_chip/stack_op_chip.rs
@@ -3,8 +3,9 @@
 use crate::chips::memory_chip::MEM_CHIP_WIDTH;
 use crate::chips::utilities::*;
 use crate::witness::rw_operations::{ConvertedRWOperation, RW};
+use crate::witness::CircuitConfig;
 use halo2_proofs::arithmetic::FieldExt;
-use halo2_proofs::circuit::{AssignedCell, Chip, Region};
+use halo2_proofs::circuit::{AssignedCell, Chip, Layouter, Region};
 use halo2_proofs::plonk::{
     Advice, Column, ConstraintSystem, Error, Expression, Selector, TableColumn,
 };
@@ -232,7 +233,7 @@ impl<F: FieldExt> StackOpChip<F> {
     }
 
     // assign each cell of the stack operation, return assigned cell for counter
-    pub fn assign(
+    pub fn assign_cell(
         &self,
         region: &mut Region<'_, F>,
         offset: usize,
@@ -310,5 +311,69 @@ impl<F: FieldExt> StackOpChip<F> {
         }
 
         Ok(assigned)
+    }
+
+    pub fn assign(
+        &self,
+        layouter: &mut impl Layouter<F>,
+        _circuit_config: &CircuitConfig,
+        stack_ops: Vec<ConvertedRWOperation<F>>,
+        stack_ops_num: usize,
+    ) -> Option<AssignedCell<F, F>> {
+        let mut last_stack_counter: Option<AssignedCell<F, F>> = None;
+
+        if !stack_ops.is_empty() || stack_ops_num > 0 {
+            layouter
+                .assign_region(
+                    || "stack operations",
+                    |mut region: Region<'_, F>| {
+                        let mut counter = 0;
+                        for (index, op) in stack_ops.iter().enumerate() {
+                            counter = index + 1;
+                            let assigned_counter = if index == 0 {
+                                self.config.s_first_stack_op.enable(&mut region, index)?;
+                                self.assign_cell(&mut region, index, op, counter, false)?
+                            } else {
+                                self.config.s_stack_op.enable(&mut region, index)?;
+                                self.assign_cell(&mut region, index, op, counter, false)?
+                            };
+                            if counter == stack_ops.len() {
+                                last_stack_counter = Some(assigned_counter);
+                            }
+                        }
+
+                        // If the number of stack ops is less than stack_ops_num set by user, fill with
+                        // empty op. This happened when the execution path is not fixed, for example,
+                        // if there is loop in the code.
+                        if stack_ops.len() < stack_ops_num {
+                            for index in stack_ops.len()..stack_ops_num {
+                                let assigned_counter = if index == 0 {
+                                    self.config.s_first_stack_op.enable(&mut region, index)?;
+                                    self.assign_cell(
+                                        &mut region,
+                                        index,
+                                        &ConvertedRWOperation::empty(),
+                                        counter,
+                                        true,
+                                    )?
+                                } else {
+                                    self.config.s_stack_op.enable(&mut region, index)?;
+                                    self.assign_cell(
+                                        &mut region,
+                                        index,
+                                        &ConvertedRWOperation::empty(),
+                                        counter,
+                                        true,
+                                    )?
+                                };
+                                last_stack_counter = Some(assigned_counter);
+                            }
+                        }
+                        Ok(())
+                    },
+                )
+                .ok()?;
+        }
+        last_stack_counter
     }
 }
